Additional functions (point modifier and inventory)
I added additional functionality to the password-checking function, which displays the amount of points gained, allows for multiple attempts (and thus less points each attempt), and presents a different message once all attempts are up. I added a second function like this for a second puzzle. I also added a function representing when a user would obtain a key or prize that can either be saved to the inventory or be used to unlock the next level